Split parsing and formatting code into their own files
