+++ Exanite.Core/Numerics/Fixed128.Format.cs
@@ -1,0 +1,18 @@
+using System;
+
+namespace Exanite.Core.Numerics;
+
+public partial struct Fixed128
+{
+    public string ToString(string? format, IFormatProvider? formatProvider)
+    {
+        return ((decimal)Raw / OneRaw).ToString(format, formatProvider);
+    }
+
+    public override string ToString() => ToString(null, null);
+
+    public bool TryFormat(Span<char> destination, out int charsWritten, ReadOnlySpan<char> format, IFormatProvider? provider)
+    {
+        return ((decimal)Raw / OneRaw).TryFormat(destination, out charsWritten, format, provider);
+    }
+}
