+++ Exanite.Core/Numerics/Fixed128.cs
@@ -341,185 +341,6 @@ public readonly partial struct Fixed128 :
     public static Fixed128 MinMagnitude(Fixed128 x, Fixed128 y) => Abs(x) < Abs(y) ? x : y;
     public static Fixed128 MinMagnitudeNumber(Fixed128 x, Fixed128 y) => MinMagnitude(x, y);
 
-    // Formatting
-    public string ToString(string? format, IFormatProvider? formatProvider)
-    {
-        return ((decimal)Raw / OneRaw).ToString(format, formatProvider);
-    }
-
-    public override string ToString() => ToString(null, null);
-
-    public bool TryFormat(Span<char> destination, out int charsWritten, ReadOnlySpan<char> format, IFormatProvider? provider)
-    {
-        return ((decimal)Raw / OneRaw).TryFormat(destination, out charsWritten, format, provider);
-    }
-
-    // Parsing
-    public static bool TryParse(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider? provider, out Fixed128 result)
-    {
-        // Intentionally excluded values:
-        // AllowExponent
-        // AllowCurrencySymbol
-        const NumberStyles allowedStyles = 0
-            | NumberStyles.AllowLeadingWhite
-            | NumberStyles.AllowTrailingWhite
-            | NumberStyles.AllowLeadingSign
-            | NumberStyles.AllowTrailingSign
-            | NumberStyles.AllowParentheses
-            | NumberStyles.AllowDecimalPoint
-            | NumberStyles.AllowThousands
-            | NumberStyles.AllowHexSpecifier
-            | NumberStyles.AllowBinarySpecifier;
-
-        if ((style & ~allowedStyles) != 0)
-        {
-            result = default;
-            return false;
-        }
-
-        // If hex or binary is requested, then just parse it as an Int128 and shift up
-        if ((style & (NumberStyles.AllowHexSpecifier | NumberStyles.AllowBinarySpecifier)) != 0)
-        {
-            if (Int128.TryParse(s, style, provider, out var value))
-            {
-                if (value > (Int128)MaxValue || value < (Int128)MinValue)
-                {
-                    result = default;
-                    return false;
-                }
-
-                result = new Fixed128(value << Shift);
-                return true;
-            }
-        }
-
-        // Handle whitespace
-        if ((style & NumberStyles.AllowLeadingWhite) != 0)
-        {
-            s = s.TrimStart();
-        }
-
-        if ((style & NumberStyles.AllowTrailingWhite) != 0)
-        {
-            s = s.TrimEnd();
-        }
-
-        if (s.IsEmpty)
-        {
-            result = default;
-            return false;
-        }
-
-        // Get format info
-        var formatInfo = NumberFormatInfo.GetInstance(provider);
-
-        // Handle signs
-        var isNegative = false;
-        if ((style & NumberStyles.AllowParentheses) != 0 && s.StartsWith("(") && s.EndsWith(")"))
-        {
-            isNegative = true;
-            s = s[1..^1];
-        }
-        else if ((style & NumberStyles.AllowLeadingSign) != 0 && s.StartsWith(formatInfo.NegativeSign))
-        {
-            isNegative = true;
-            s = s[formatInfo.NegativeSign.Length..];
-        }
-        else if ((style & NumberStyles.AllowLeadingSign) != 0 && s.StartsWith(formatInfo.PositiveSign))
-        {
-            s = s[formatInfo.PositiveSign.Length..];
-        }
-
-        // Split into integral and fractional parts
-        ReadOnlySpan<char> integralText;
-        ReadOnlySpan<char> fractionalText;
-
-        var decimalSeparator = formatInfo.NumberDecimalSeparator;
-        var decimalIndex = s.IndexOf(decimalSeparator);
-        if (decimalIndex >= 0)
-        {
-            if ((style & NumberStyles.AllowDecimalPoint) == 0)
-            {
-                result = default;
-                return false;
-            }
-
-            integralText = s[..decimalIndex];
-            fractionalText = s[(decimalIndex + decimalSeparator.Length)..];
-        }
-        else
-        {
-            integralText = s;
-            fractionalText = ReadOnlySpan<char>.Empty;
-        }
-
-        // Parse integral portion
-        // Group separators can appear in the integral portion, but not the fractional portion
-        var integralStyle = style & NumberStyles.AllowThousands;
-        if (!Int128.TryParse(integralText, integralStyle, provider, out var integralValue))
-        {
-            result = default;
-            return false;
-        }
-
-        var resultRaw = integralValue << Shift;
-
-        // Parse fractional portion
-        var divisor = 10L;
-        var tenRaw = OneRaw * 10;
-        foreach (var c in fractionalText)
-        {
-            if (!char.IsDigit(c))
-            {
-                result = default;
-                return false;
-            }
-
-            var digit = c - '0';
-
-            // Round up by adding the last bit
-            // This adds a tiny bit more precision
-            var fraction = ((Int128)(OneRaw << 1) * digit) / divisor;
-            fraction += fraction & 1;
-            fraction >>= 1;
-
-            resultRaw += fraction;
-
-            divisor *= 10;
-            if (divisor > tenRaw)
-            {
-                break;
-            }
-        }
-
-        // Apply negative sign
-        if (isNegative)
-        {
-            resultRaw = -resultRaw;
-        }
-
-        result = new Fixed128(resultRaw);
-        return true;
-    }
-
-    public static bool TryParse(ReadOnlySpan<char> s, IFormatProvider? provider, out Fixed128 result) => TryParse(s, NumberStyles.Float | NumberStyles.AllowThousands, provider, out result);
-    public static bool TryParse([NotNullWhen(true)] string? s, IFormatProvider? provider, out Fixed128 result) => TryParse((ReadOnlySpan<char>)s, provider, out result);
-    public static bool TryParse([NotNullWhen(true)] string? s, NumberStyles style, IFormatProvider? provider, out Fixed128 result) => TryParse((ReadOnlySpan<char>)s, style, provider, out result);
-
-    public static Fixed128 Parse(ReadOnlySpan<char> s, NumberStyles style, IFormatProvider? provider)
-    {
-        if (!TryParse(s, style, provider, out var result))
-        {
-            throw new FormatException($"Input string '{s.ToString()}' was not in a correct format.");
-        }
-
-        return result;
-    }
-
-    public static Fixed128 Parse(ReadOnlySpan<char> s, IFormatProvider? provider) => Parse(s, NumberStyles.Float | NumberStyles.AllowThousands, provider);
-    public static Fixed128 Parse(string s, IFormatProvider? provider) => Parse((ReadOnlySpan<char>)s, provider);
-    public static Fixed128 Parse(string s, NumberStyles style, IFormatProvider? provider) => Parse((ReadOnlySpan<char>)s, style, provider);
-
     // ToFromRaw
     public Int128 ToRaw()
     {
